SurveyMixin introduced to manage workflow surveys
