+++ tower_cli/models/__init__.py
@@ -18,6 +18,6 @@
 from __future__ import absolute_import, unicode_literals
 
 from tower_cli.models.base import Resource, MonitorableResource, ExeResource, \
-                                  BaseResource
+                                  BaseResource, SurveyResource
 from tower_cli.models.fields import Field
 from tower_cli.utils.types import File
